Saving wip. Wrapping CQN in its on CqnConn struct
